#146
- Update JasperReports Library 7.0.1
- Updated the developers instructions for setting up the environment and testing a new release
- Upgraded all .jrxml files to version 7.0.0
