+++ src/main/webapp/printerDiagnostics.jsp
@@ -1,6 +1,6 @@
 <%@page import="de.oc.print.PrinterUtilities"%>
 <%@page import="de.oc.integration.jasper.webapp.AppConfig"%>
-<%@page import="org.apache.commons.lang.StringEscapeUtils" %>
+<%@page import="org.apache.commons.text.StringEscapeUtils" %>
 
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
 	pageEncoding="ISO-8859-1" import="java.io.File"
@@ -52,7 +52,7 @@
 		</thead>
 		<%
 			for (int i = 0; i < printers.size(); i++) {
-				String printerName = StringEscapeUtils.escapeHtml(((PrintService) printers.get(i)).getName());
+				String printerName = StringEscapeUtils.escapeHtml4(((PrintService) printers.get(i)).getName());
 		%>
 		<tr>
 			<td><a href="#<%=printerName%>"><%=printerName%></a></td>
@@ -70,7 +70,7 @@
 	<%
 		for (int i = 0; i < printers.size(); i++) {
 			PrintService printService = (PrintService) printers.get(i);
-			String printerName = StringEscapeUtils.escapeHtml(printService.getName());
+			String printerName = StringEscapeUtils.escapeHtml4(printService.getName());
 	%>
 	<h3>
 		<a name="<%=printerName%>"></a><jsp:expression>printerName</jsp:expression></h3>
@@ -92,7 +92,7 @@
 				for (int k = 0; k < attributes.length; k++) {
 		%>
 		<tr>
-			<td><%=StringEscapeUtils.escapeHtml(attributes[k].getName())%></td>
+			<td><%=StringEscapeUtils.escapeHtml4(attributes[k].getName())%></td>
 			<td><jsp:expression>attributes[k]</jsp:expression></td>
 		</tr>
 		<%
@@ -120,7 +120,7 @@
 				for (int k = 0; k < attributeCategories.length; k++) {
 		%>
 		<tr>
-			<td><%=StringEscapeUtils.escapeHtml(attributeCategories[k].getName())%></td>
+			<td><%=StringEscapeUtils.escapeHtml4(attributeCategories[k].getName())%></td>
 			<td>
 				<%
 					@SuppressWarnings("unchecked")
@@ -143,12 +143,12 @@
 										Object[] values = (Object[]) valueObject;
 										for (int j = 0; j < values.length; j++) {
 											if (values[j] != null) {
-												out.println(StringEscapeUtils.escapeHtml(values[j].toString()));
+												out.println(StringEscapeUtils.escapeHtml4(values[j].toString()));
 												out.println("<br>");
 											}
 										}
 									} else {
-										out.println(StringEscapeUtils.escapeHtml(valueObject.toString()));
+										out.println(StringEscapeUtils.escapeHtml4(valueObject.toString()));
 										out.println("<br>");
 									}
