+++ src/reports/demo/test_images.jrxml
@@ -1,87 +1,59 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Created with Jaspersoft Studio version 6.14.0.final using JasperReports Library version 6.14.0-2ab0d8625be255bf609c78e1181801213e51db8f  -->
-<jasperReport xmlns="http://jasperreports.sourceforge.net/jasperreports" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://jasperreports.sourceforge.net/jasperreports http://jasperreports.sourceforge.net/xsd/jasperreport.xsd" name="test2" language="groovy" pageWidth="595" pageHeight="842" whenNoDataType="AllSectionsNoDetail" columnWidth="535" leftMargin="30" rightMargin="30" topMargin="20" bottomMargin="20" uuid="c2c7f585-dca1-436e-948b-83b710d4542f">
+<jasperReport name="test2" language="groovy" pageWidth="595" pageHeight="842" whenNoDataType="AllSectionsNoDetail" columnWidth="535" leftMargin="30" rightMargin="30" topMargin="20" bottomMargin="20" uuid="c2c7f585-dca1-436e-948b-83b710d4542f">
 	<property name="ireport.scriptlethandling" value="0"/>
 	<property name="ireport.encoding" value="UTF-8"/>
 	<property name="com.jaspersoft.studio.data.defaultdataadapter" value="jri_test@vm1.xml"/>
-	<import value="net.sf.jasperreports.engine.*"/>
-	<import value="java.util.*"/>
-	<import value="net.sf.jasperreports.engine.data.*"/>
+	<import>net.sf.jasperreports.engine.*</import>
+	<import>java.util.*</import>
+	<import>net.sf.jasperreports.engine.data.*</import>
 	<parameter name="BaseDir" class="java.lang.String">
 		<defaultValueExpression><![CDATA["/Users/daust/Dropbox/50.Projects/github/JasperReportsIntegration/src/reports/demo2"]]></defaultValueExpression>
 	</parameter>
-	<background>
-		<band splitType="Stretch"/>
-	</background>
-	<title>
-		<band height="200" splitType="Stretch">
-			<staticText>
-				<reportElement key="staticText-1" x="0" y="80" width="380" height="27" uuid="922c355f-4331-4bc6-82e4-2facc43fafe6"/>
-				<box>
-					<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-				</box>
-				<textElement textAlignment="Left">
-					<font fontName="Arial" size="17"/>
-				</textElement>
-				<text><![CDATA[Test with image (relative path)]]></text>
-			</staticText>
-			<image scaleImage="RetainShape">
-				<reportElement key="image-1" x="399" y="12" width="115" height="50" uuid="0af74598-05bf-47e1-ad5a-52a16f93be2f"/>
-				<box>
-					<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-				</box>
-				<imageExpression><![CDATA[$P{BaseDir} + "/opal_logo_50px_hoch.jpg"]]></imageExpression>
-			</image>
-			<image scaleImage="RetainShape">
-				<reportElement key="image-1" x="399" y="80" width="115" height="50" uuid="f123d6e5-9a93-4a39-ba56-95e16f1fd25e"/>
-				<box>
-					<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-				</box>
-				<imageExpression><![CDATA["opal_logo_50px_hoch.jpg"]]></imageExpression>
-			</image>
-			<staticText>
-				<reportElement key="staticText-1" x="0" y="12" width="380" height="28" uuid="a9afffea-f699-4e76-ac0f-4f1da9a1505b"/>
-				<box>
-					<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-					<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
-				</box>
-				<textElement textAlignment="Left">
-					<font fontName="Arial" size="17"/>
-				</textElement>
-				<text><![CDATA[Test with image (absolute path using BaseDir)]]></text>
-			</staticText>
-		</band>
+	<background splitType="Stretch"/>
+	<title height="200" splitType="Stretch">
+		<element kind="staticText" uuid="922c355f-4331-4bc6-82e4-2facc43fafe6" key="staticText-1" x="0" y="80" width="380" height="27" fontName="Arial" fontSize="17.0" hTextAlign="Left">
+			<text><![CDATA[Test with image (relative path)]]></text>
+			<box>
+				<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+			</box>
+		</element>
+		<element kind="image" uuid="0af74598-05bf-47e1-ad5a-52a16f93be2f" key="image-1" x="399" y="12" width="115" height="50" scaleImage="RetainShape">
+			<expression><![CDATA[$P{BaseDir} + "/opal_logo_50px_hoch.jpg"]]></expression>
+			<box>
+				<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+			</box>
+		</element>
+		<element kind="image" uuid="f123d6e5-9a93-4a39-ba56-95e16f1fd25e" key="image-1" x="399" y="80" width="115" height="50" scaleImage="RetainShape">
+			<expression><![CDATA["opal_logo_50px_hoch.jpg"]]></expression>
+			<box>
+				<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+			</box>
+		</element>
+		<element kind="staticText" uuid="a9afffea-f699-4e76-ac0f-4f1da9a1505b" key="staticText-1" x="0" y="12" width="380" height="28" fontName="Arial" fontSize="17.0" hTextAlign="Left">
+			<text><![CDATA[Test with image (absolute path using BaseDir)]]></text>
+			<box>
+				<topPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<leftPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<bottomPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+				<rightPen lineWidth="0.0" lineStyle="Solid" lineColor="#000000"/>
+			</box>
+		</element>
 	</title>
-	<pageHeader>
-		<band height="50" splitType="Stretch"/>
-	</pageHeader>
-	<columnHeader>
-		<band height="30" splitType="Stretch"/>
-	</columnHeader>
+	<pageHeader height="50" splitType="Stretch"/>
+	<columnHeader height="30" splitType="Stretch"/>
 	<detail>
 		<band height="100" splitType="Stretch"/>
 	</detail>
-	<columnFooter>
-		<band height="30" splitType="Stretch"/>
-	</columnFooter>
-	<pageFooter>
-		<band height="50" splitType="Stretch"/>
-	</pageFooter>
-	<lastPageFooter>
-		<band height="50" splitType="Stretch"/>
-	</lastPageFooter>
-	<summary>
-		<band height="50" splitType="Stretch"/>
-	</summary>
+	<columnFooter height="30" splitType="Stretch"/>
+	<pageFooter height="50" splitType="Stretch"/>
+	<lastPageFooter height="50" splitType="Stretch"/>
+	<summary height="50" splitType="Stretch"/>
 </jasperReport>
